+++ content_manager_wrapper.go
@@ -71,6 +71,9 @@ type CMAssists struct {
 type CMContent struct {
 	Cars  map[string]ContentURL `json:"cars"`
 	Track ContentURL            `json:"track"`
+
+	// @TODO not functional
+	Password bool `json:"password"`
 }
 
 type ContentURL struct {
@@ -118,7 +121,7 @@ func NewContentManagerWrapper(store Store, carManager *CarManager, trackManager 
 	}
 }
 
-func (cmw *ContentManagerWrapper) NewCMContent(cars []string, trackName string) (*CMContent, error) {
+func (cmw *ContentManagerWrapper) NewCMContent(cars []string, trackName string, requirePassword bool) (*CMContent, error) {
 	carsMap := make(map[string]ContentURL)
 	var trackDownload string
 
@@ -146,6 +149,7 @@ func (cmw *ContentManagerWrapper) NewCMContent(cars []string, trackName string) 
 		Track: ContentURL{
 			URL: trackDownload,
 		},
+		Password: requirePassword,
 	}, nil
 }
 
@@ -409,7 +413,8 @@ func (cmw *ContentManagerWrapper) buildContentManagerDetails(guid string) (*Cont
 
 	description += cmw.description
 
-	cmContent, err := cmw.NewCMContent(sessionInfo.Cars, race.Track)
+	// @TODO ContentManagerWrapperContentRequiresPassword from config_ini.go
+	cmContent, err := cmw.NewCMContent(sessionInfo.Cars, race.Track, false)
 
 	if err != nil {
 		logrus.Errorf("Couldn't attach content download URL(s) through CM Wrapper!")